requirements.txt
# Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# API Documentation
drf-yasg==1.21.7

# WebSockets
channels==4.0.0
channels-redis==4.1.0

# Async support
daphne==4.0.0

# Image handling
Pillow==10.2.0

# Email
django-anymail==10.2

# SMS (Twilio)
twilio==8.11.1

# Payment Gateway (Stripe)
stripe==7.11.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code quality
flake8==7.0.0
black==23.12.1

# Security
django-ratelimit==4.1.0
django-defender==0.9.5

# Monitoring
sentry-sdk==1.39.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
